Get stuff compiling again after splitting post processors
