+++ grakn-engine/src/main/java/ai/grakn/engine/controller/GraqlController.java
@@ -108,8 +108,7 @@ public class GraqlController {
     private final Timer executeExplanation;
 
     public GraqlController(
-            EngineGraknTxFactory factory, Service spark,
-            PostProcessor postProcessor, Printer printer, MetricRegistry metricRegistry
+            EngineGraknTxFactory factory, Service spark, PostProcessor postProcessor, Printer printer, MetricRegistry metricRegistry
     ) {
         this.factory = factory;
         this.postProcessor = postProcessor;
